Rename fasteval -> bencheval
