+++ bencheval/pyproject.toml
@@ -3,7 +3,7 @@ requires = ["setuptools>=69", "wheel"]
 build-backend = "setuptools.build_meta"
 
 [project]
-name = "fasteval"
+name = "bencheval"
 version = "0.0.1"
 requires-python = ">=3.10"
 license = { text = "Apache-2.0" }
@@ -23,11 +23,11 @@ dependencies = [
 
 [tool.setuptools.packages.find]
 where = ["."]
-include = ["fasteval*"]
+include = ["bencheval*"]
 
 #[tool.setuptools]
 #package-dir = {}
-#packages = ["fasteval"]
+#packages = ["bencheval"]
 
 [project.urls]
 Homepage = "https://github.com/autogluon/tabarena"
