+++ examples/plots/run_generate_main_leaderboard.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 from pathlib import Path
 
 from tabarena.nips2025_utils.tabarena_context import TabArenaContext
-from fasteval.website_format import format_leaderboard
+from bencheval.website_format import format_leaderboard
 
 if __name__ == "__main__":
     save_path = "output_leaderboard"  # folder to save all figures and tables
