+++ tst/test_elo_utils.py
@@ -1,5 +1,5 @@
 import pytest
-from fasteval.elo_utils import EloHelper
+from bencheval.elo_utils import EloHelper
 import pandas as pd
